bot/requirements.txt
requests
discord.py
pymongo